Fix merge conflict with CMSSW_16_1_X

+++ RecoTracker/PixelSeeding/plugins/alpaka/BrokenLineFit.dev.cc
@@ -63,7 +63,7 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE {
         }
         // get it from the ntuple container (one to one to helix)
         auto tkid = *(tupleMultiplicity->begin(nHitsL) + tuple_idx);
-        ALPAKA_ASSERT_ACC(static_cast<int>(tkid) < foundNtuplets->nOnes());
+        ALPAKA_ASSERT_ACC(tkid < foundNtuplets->nOnes());
 
         ptkids[local_idx] = tkid;
 
@@ -206,7 +206,7 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE {
           break;
         auto tkid = ptkids[local_idx];
 
-        ALPAKA_ASSERT_ACC(int(tkid) < tupleMultiplicity->capacity());
+        ALPAKA_ASSERT_ACC(tkid < tupleMultiplicity->capacity());
 
         riemannFit::Map3xNd<N> hits(phits + local_idx);
         riemannFit::Map4d fast_fit(pfast_fit + local_idx);
